move from karma to web-test-runner
to fix coverage reporting
