+++ packages/rtcstats-js/package.json
@@ -6,19 +6,12 @@
   "type": "module",
   "devDependencies": {
     "@puppeteer/browsers": "^2.10.9",
-    "jest": "^30.1.3",
-    "karma": "^6.4.4",
-    "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^3.2.0",
-    "karma-coverage": "^2.2.1",
-    "karma-firefox-launcher": "^2.1.3",
-    "karma-mocha": "^2.0.1",
-    "karma-mocha-reporter": "^2.2.5",
-    "karma-webpack": "^5.0.1"
+    "@esm-bundle/chai": "^4.3.4-fix.0",
+    "@web/test-runner": "^0.20.2"
   },
   "scripts": {
     "lint": "eslint *.js test/",
-    "coverage": "c8 karma start test/karma.conf.js",
+    "coverage": "web-test-runner --config test/web-test-runner.config.mjs",
     "test": "npm run lint && npm run coverage"
   },
   "repository": {
